+++ src/shinestacker/config/gui_constants.py
@@ -33,8 +33,9 @@ class _GuiConstants:
     }
 
     DEFAULT_MIN_MOUSE_STEP_BRUSH_FRACTION = 0.25
-    DEFAULT_PAINT_REFRESH_TIME = 50  # milliseconds
-    DEFAULT_DISPLAY_REFRESH_TIME = 50  # milliseconds
+    DEFAULT_PAINT_REFRESH_TIME = 50  # ms
+    DEFAULT_DISPLAY_REFRESH_TIME = 50  # ms
+    DEFAULT_CURSOR_UPDATE_TIME = 16  # ms
 
     THUMB_WIDTH = 120  # px
     THUMB_HEIGHT = 80  # px